kuring-69 [수정] 학과 로직, 테스트 코드를 :data:department로 옮김
