+++ app/src/main/java/com/ku_stacks/ku_ring/di/NetworkModule.kt
@@ -24,10 +24,4 @@ object NetworkModule {
     fun provideSendbirdClient(sendbirdService: SendbirdService): SendbirdClient {
         return SendbirdClient(sendbirdService)
     }
-
-    @Provides
-    @Singleton
-    fun provideDepartmentService(@Named("Default") retrofit: Retrofit): DepartmentService {
-        return retrofit.create(DepartmentService::class.java)
-    }
 }
