+++ app/src/main/java/com/ku_stacks/ku_ring/di/RepositoryModule.kt
@@ -1,9 +1,6 @@
 package com.ku_stacks.ku_ring.di
 
-import com.ku_stacks.ku_ring.data.api.DepartmentClient
 import com.ku_stacks.ku_ring.data.api.SendbirdClient
-import com.ku_stacks.ku_ring.data.db.DepartmentDao
-import com.ku_stacks.ku_ring.preferences.PreferenceUtil
 import com.ku_stacks.ku_ring.repository.*
 import dagger.Module
 import dagger.Provides
@@ -17,26 +14,10 @@ object RepositoryModule {
 
     @Provides
     @Singleton
-    fun provideSubscribeRepository(
-        departmentClient: DepartmentClient,
-        pref: PreferenceUtil
-    ): SubscribeRepository {
-        return SubscribeRepositoryImpl(departmentClient, pref)
-    }
-
-    @Provides
-    @Singleton
     fun provideSendbirdRepository(
         sendbirdClient: SendbirdClient
     ): SendbirdRepository {
         return SendbirdRepositoryImpl(sendbirdClient)
     }
-
-    @Provides
-    @Singleton
-    fun provideDepartmentRepository(
-        departmentDao: DepartmentDao,
-        departmentClient: DepartmentClient
-    ): DepartmentRepository = DepartmentRepositoryImpl(departmentDao, departmentClient)
 
 }
