+++ data/department/src/main/java/com/ku_stacks/ku_ring/department/local/DepartmentDatabase.kt
@@ -1,0 +1,13 @@
+package com.ku_stacks.ku_ring.department.local
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(
+    entities = [DepartmentEntity::class],
+    version = 6,
+    exportSchema = false,
+)
+abstract class DepartmentDatabase: RoomDatabase() {
+    abstract fun departmentDao(): DepartmentDao
+}
