+++ data/department/src/main/java/com/ku_stacks/ku_ring/department/mapper/DomainToEntity.kt
@@ -1,6 +1,6 @@
-package com.ku_stacks.ku_ring.data.mapper
+package com.ku_stacks.ku_ring.department.mapper
 
-import com.ku_stacks.ku_ring.data.db.DepartmentEntity
+import com.ku_stacks.ku_ring.department.local.DepartmentEntity
 import com.ku_stacks.ku_ring.domain.Department
 
 fun List<Department>.toEntityList(): List<DepartmentEntity> = map { it.toEntity() }
