+++ data/department/src/main/java/com/ku_stacks/ku_ring/department/mapper/EntityToDomain.kt
@@ -1,8 +1,7 @@
-package com.ku_stacks.ku_ring.data.mapper
+package com.ku_stacks.ku_ring.department.mapper
 
-import com.ku_stacks.ku_ring.data.db.DepartmentEntity
+import com.ku_stacks.ku_ring.department.local.DepartmentEntity
 import com.ku_stacks.ku_ring.domain.Department
-
 
 fun List<DepartmentEntity>.toDepartmentList() = map { it.toDepartment() }
