+++ data/department/src/main/java/com/ku_stacks/ku_ring/department/mapper/ResponseToDomain.kt
@@ -1,6 +1,6 @@
-package com.ku_stacks.ku_ring.data.mapper
+package com.ku_stacks.ku_ring.department.mapper
 
-import com.ku_stacks.ku_ring.data.api.response.DepartmentResponse
+import com.ku_stacks.ku_ring.department.remote.response.DepartmentResponse
 import com.ku_stacks.ku_ring.domain.Department
 
 fun DepartmentResponse.toDepartment() = Department(
