+++ data/department/src/main/java/com/ku_stacks/ku_ring/department/remote/response/DepartmentResponse.kt
@@ -1,4 +1,4 @@
-package com.ku_stacks.ku_ring.data.api.response
+package com.ku_stacks.ku_ring.department.remote.response
 
 import com.google.gson.annotations.SerializedName
